Removing cyclic dependency on MiniAccumuloCluster by moving tests to test module
