+++ core/src/main/java/org/apache/accumulo/core/iterators/user/CfCqSliceFilter.java
@@ -1,3 +1,19 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.accumulo.core.iterators.user;
 
 import org.apache.accumulo.core.data.Key;
@@ -13,71 +29,68 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.util.Map;
 
+public class CfCqSliceFilter extends Filter implements OptionDescriber {
 
-public class ColumnSliceFilter extends Filter implements OptionDescriber {
+  private static final Logger log = LoggerFactory.getLogger(ColumnSliceFilter.class);
 
-    private static final Logger log = LoggerFactory.getLogger(ColumnSliceFilter.class);
+  private CfCqSliceOpts cso;
 
-    private ColumnSliceOpts cso;
+  @Override
+  public void init(SortedKeyValueIterator<Key,Value> source, Map<String,String> options, IteratorEnvironment env) throws IOException {
+    super.init(source, options, env);
+    cso = new CfCqSliceOpts(options);
+  }
 
-    @Override
-    public void init(SortedKeyValueIterator<Key, Value> source, Map<String, String> options, IteratorEnvironment env) throws IOException {
-        super.init(source, options, env);
-        cso = new ColumnSliceOpts(options);
+  @Override
+  public boolean accept(Key k, Value v) {
+    PartialKey inSlice = isKeyInSlice(k);
+    return inSlice == PartialKey.ROW_COLFAM_COLQUAL;
+  }
+
+  @Override
+  public SortedKeyValueIterator<Key,Value> deepCopy(IteratorEnvironment env) {
+    CfCqSliceFilter o = (CfCqSliceFilter) super.deepCopy(env);
+    o.cso = new CfCqSliceOpts(cso);
+    return o;
+  }
+
+  private PartialKey isKeyInSlice(Key k) {
+    if (cso.minCf.getLength() > 0) {
+      int minCfComp = k.compareColumnFamily(cso.minCf);
+      if (minCfComp < 0 || (minCfComp == 0 && !cso.minInclusive)) {
+        return PartialKey.ROW;
+      }
     }
-
-    @Override
-    public boolean accept(Key k, Value v) {
-        PartialKey inSlice = isKeyInSlice(k);
-        return inSlice == PartialKey.ROW_COLFAM_COLQUAL;
+    if (cso.maxCf.getLength() > 0) {
+      int maxCfComp = k.compareColumnFamily(cso.maxCf);
+      if (maxCfComp > 0 || (maxCfComp == 0 && !cso.maxInclusive)) {
+        return PartialKey.ROW;
+      }
     }
-
-    @Override
-    public SortedKeyValueIterator<Key, Value> deepCopy(IteratorEnvironment env) {
-        ColumnSliceFilter o = (ColumnSliceFilter)super.deepCopy(env);
-        o.cso = new ColumnSliceOpts(cso);
-        return o;
+    // k.colfam is in the "slice".
+    if (cso.minCq.getLength() > 0) {
+      int minCqComp = k.compareColumnQualifier(cso.minCq);
+      if (minCqComp < 0 || (minCqComp == 0 && !cso.minInclusive)) {
+        return PartialKey.ROW_COLFAM;
+      }
     }
-
-    private PartialKey isKeyInSlice(Key k) {
-        if (cso.minCf.getLength() > 0) {
-            int minCfComp = k.compareColumnFamily(cso.minCf);
-            if (minCfComp < 0 || (minCfComp == 0 && !cso.minInclusive)) {
-                log.info("cf " + k.getColumnFamily() + " < " + cso.minCf + ", isKeyInSlice == ROW");
-                return PartialKey.ROW;
-            }
-        }
-        if (cso.maxCf.getLength() > 0) {
-            int maxCfComp = k.compareColumnFamily(cso.maxCf);
-            if (maxCfComp > 0 || (maxCfComp == 0 && !cso.maxInclusive)) {
-                log.info("cf " + k.getColumnFamily() + " > " + cso.maxCf + ", isKeyInSlice == ROW");
-                return PartialKey.ROW;
-            }
-        }
-        // k.colfam is in the "slice".
-        if (cso.minCq.getLength() > 0) {
-            int minCqComp = k.compareColumnQualifier(cso.minCq);
-            if (minCqComp < 0 || (minCqComp == 0 && !cso.minInclusive)) {
-                return PartialKey.ROW_COLFAM;
-            }
-        }
-        if (cso.maxCq.getLength() > 0) {
-            int maxCqComp = k.compareColumnQualifier(cso.maxCq);
-            if (maxCqComp > 0 || (maxCqComp == 0 && !cso.maxInclusive)) {
-                return PartialKey.ROW_COLFAM;
-            }
-        }
-        // k.colqual is in the slice.
-        return PartialKey.ROW_COLFAM_COLQUAL;
+    if (cso.maxCq.getLength() > 0) {
+      int maxCqComp = k.compareColumnQualifier(cso.maxCq);
+      if (maxCqComp > 0 || (maxCqComp == 0 && !cso.maxInclusive)) {
+        return PartialKey.ROW_COLFAM;
+      }
     }
+    // k.colqual is in the slice.
+    return PartialKey.ROW_COLFAM_COLQUAL;
+  }
 
-    @Override
-    public IteratorOptions describeOptions() {
-        return cso.describeOptions();
-    }
+  @Override
+  public IteratorOptions describeOptions() {
+    return cso.describeOptions();
+  }
 
-    @Override
-    public boolean validateOptions(Map<String, String> options) {
-        return cso.validateOptions(options);
-    }
+  @Override
+  public boolean validateOptions(Map<String,String> options) {
+    return cso.validateOptions(options);
+  }
 }
