--- core/src/test/java/org/apache/accumulo/core/iterators/user/TestCfCqSliceFilter.java
@@ -1,8 +1,0 @@
-package org.apache.accumulo.core.iterators.user;
-
-public class TestColumnSliceFilter extends TestColumnSlice {
-    @Override
-    protected Class getFilterClass() {
-        return ColumnSliceFilter.class;
-    }
-}
