--- core/src/test/java/org/apache/accumulo/core/iterators/user/TestCfCqSliceSeekingFilter.java
@@ -1,8 +1,0 @@
-package org.apache.accumulo.core.iterators.user;
-
-public class TestColumnSliceSeekingFilter extends TestColumnSlice {
-    @Override
-    protected Class getFilterClass() {
-        return ColumnSliceSeekingFilter.class;
-    }
-}
